+++ common/TracySystem.hpp
@@ -10,20 +10,20 @@ namespace tracy
 
 namespace detail
 {
-TRACY_API uint64_t GetThreadHandleImpl();
+TRACY_API uint32_t GetThreadHandleImpl();
 }
 
 #ifdef TRACY_ENABLE
-TRACY_API uint64_t GetThreadHandle();
+TRACY_API uint32_t GetThreadHandle();
 #else
-static inline uint64_t GetThreadHandle()
+static inline uint32_t GetThreadHandle()
 {
     return detail::GetThreadHandleImpl();
 }
 #endif
 
 TRACY_API void SetThreadName( const char* name );
-TRACY_API const char* GetThreadName( uint64_t id );
+TRACY_API const char* GetThreadName( uint32_t id );
 
 TRACY_API const char* GetEnvVar(const char* name);
